+++ platform/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.2.1] platform
 ### Releases
 - [NPM v1.2.1] https://www.npmjs.com/package/@codat/platform/v/1.2.1 - platform
+
+## 2023-09-29 00:04:01
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.0] platform
+### Releases
+- [NPM v1.3.0] https://www.npmjs.com/package/@codat/platform/v/1.3.0 - platform